Streaming logs from worker to browser through web
Uses Redis as the intermediary - the worker writes logs to a
redis pub/sub channel. The web server subscribes to this channel.
Each time the web server receives a message, it rebroadcasts
the message on a socket.io namespace. The browser subscribes
to this namespace, with the final result being that logs from
the worker show up in the browser